+++ uploader/mock_http_test.go
@@ -6,12 +6,10 @@ import (
 
 // HttpClient mocks an HTTP client.
 type HttpClient struct {
-	DoFn      func(req *http.Request) (*http.Response, error)
-	DoInvoked bool
+	DoFn func(req *http.Request) (*http.Response, error)
 }
 
 // Do invokes the mock implementation and marks the function as invoked.
 func (c *HttpClient) Do(req *http.Request) (*http.Response, error) {
-	c.DoInvoked = true
 	return c.DoFn(req)
 }
